# public/swagger.yaml
openapi: 3.0.0
info:
  title: Lentera Hijaiyah API
  description: Dokumentasi API untuk sistem autentikasi dan manajemen akun Custom JWT.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/v1/auth/forgot-password:
    post:
      summary: Kirim email reset password
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: ahmad@example.com
      responses:
        "200":
          description: Jika email terdaftar, instruksi reset password akan dikirim
        "400":
          description: Input tidak valid / Zod Error
...
  /api/v1/learning/categories/{categoryId}/modules:
    get:
      summary: Ambil daftar modul berdasarkan kategori
      tags:
        - Learning
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
          description: ID kategori modul
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized - Token tidak valid / tidak ada
        "404":
          description: Kategori tidak ditemukan
...
  /api/v1/learning/progress:
    get:
      summary: Ambil progress belajar pengguna
      tags:
        - Learning
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized - Token tidak valid / tidak ada
    post:
      summary: Update progress belajar pengguna
      tags:
        - Learning
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - moduleId
                - isCompleted
              properties:
                moduleId:
                  type: string
                  format: uuid
                  example: 11111111-1111-1111-1111-111111111111
                isCompleted:
                  type: boolean
                  example: true
      responses:
        "200":
          description: Progress berhasil diperbarui
        "400":
          description: Input tidak valid / Zod Error
        "401":
          description: Unauthorized - Token tidak valid / tidak ada
        "404":
          description: Modul tidak ditemukan
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Masukkan JWT Token hasil return dari API Login."
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: "Cookie yang diset secara otomatis oleh browser setelah Login."